+++ lib/Table/TableAutomaticStylesStratum.ts
@@ -90,6 +90,9 @@ export default class TableAutomaticStylesStratum extends LoadableStratum(
         time: createStratumInstance(TableTimeStyleTraits, {
           timeColumn: timeColumn?.name,
           idColumns: idColumn && [idColumn.name]
+        }),
+        color: createStratumInstance(TableColorStyleTraits, {
+          legend: this._createLegendForColorStyle(-1)
         })
       });
     }
@@ -117,6 +120,9 @@ export default class TableAutomaticStylesStratum extends LoadableStratum(
 
     if (scalarColumns.length >= (hasTime ? 1 : 2)) {
       return createStratumInstance(TableStyleTraits, {
+        color: createStratumInstance(TableColorStyleTraits, {
+          legend: this._createLegendForColorStyle(-1)
+        }),
         chart: createStratumInstance(TableChartStyleTraits, {
           xAxisColumn: hasTime ? timeColumns[0].name : scalarColumns[0].name,
           lines: scalarColumns.slice(hasTime ? 0 : 1).map((column, i) =>
@@ -230,7 +236,7 @@ export class ColorStyleLegend extends LoadableStratum(LegendTraits) {
   /**
    *
    * @param catalogItem
-   * @param index index of column in catalogItem (if undefined, then default style will be used)
+   * @param index index of column in catalogItem (if -1 or undefined, then default style will be used)
    */
   constructor(
     readonly catalogItem: TableCatalogItem,
@@ -249,6 +255,7 @@ export class ColorStyleLegend extends LoadableStratum(LegendTraits) {
   @computed get tableStyle() {
     if (
       isDefined(this.index) &&
+      this.index !== -1 &&
       this.index < this.catalogItem.tableStyles.length
     )
       return this.catalogItem.tableStyles[this.index];